Build: separate files so the package builds on Windows
I encountered this when pulling in a package that uses expect
indirectly, and then tried to build.

Split OS-specific content into files separated by build tags
fixed the problem.
