+++ expect.go
@@ -33,6 +33,7 @@ line is:
 
 This package has only been tested on Linux
 */
+
 package expect
 
 import (
@@ -45,11 +46,8 @@ import (
 	"os"
 	"os/exec"
 	"regexp"
-	"runtime"
 	"syscall"
 	"time"
-
-	"github.com/kr/pty"
 )
 
 const (
@@ -74,20 +72,7 @@ var (
 	// ExpectInSize is the size of the channel between the expectReader and Expect.
 	// If you overflow this than expectReader will block.
 	ExpectInSize = 20 * 1024
-
-	// EOF is the terminal EOF character - NOT guaranteed to be correct on all
-	// systems nor if the program passed to expect connects to a different
-	// platform
-	EOF = "\004"
 )
-
-func init() {
-	// No good way to discover EOF so have to take a best guess
-	// Default (above) to control-D
-	if runtime.GOOS == "windows" {
-		EOF = "\032" // control-Z
-	}
-}
 
 // Remember: Expect.Close() will not end the process
 // you have to send it an EOF
@@ -156,45 +141,6 @@ func NewExpectProc(ctx context.Context, prog string, arg ...string) (*Expect, *e
 		return nil, nil, err
 	}
 	return exp, exp.cmd, err
-}
-
-// newExpectCommon uses @parentCtx as cancellation context and @reap to indicate
-// whether the spawned process should automatically be reaped.
-func newExpectCommon(parentCtx context.Context, reap bool, prog string, arg ...string) (exp *Expect, err error) {
-	defer func() {
-		// On an error I want to kill the process - if it was started
-		if err != nil && exp != nil && exp.cmd.Process != nil {
-			debugf("killing process %q due to error (%s)", prog, err)
-			exp.Kill()
-		}
-	}()
-
-	exp = &Expect{
-		Buffer:  new(bytes.Buffer),
-		bytesIn: make(chan byteIn, ExpectInSize),
-	}
-
-	// Create a cancelable child context for exp.cmd (go >= 1.7)
-	exp.ctx, exp.cancel = context.WithCancel(parentCtx)
-	exp.cmd = exec.CommandContext(exp.ctx, prog, arg...)
-
-	if exp.file, err = pty.Start(exp.cmd); err != nil {
-		return nil, err
-	}
-
-	// make the pty non blocking so when I read from it I dont jam up
-	if err = syscall.SetNonblock(int(exp.file.Fd()), true); err != nil {
-		return nil, err
-	}
-	exp.SetCmdOut(nil)
-
-	go exp.expectReader()
-
-	if reap {
-		go exp.expectReaper()
-	}
-
-	return exp, nil
 }
 
 // expectReaper reaps the process if it ends for any reason and saves the
